refactor: move ClinicalDataDecoder class to use_case layer
