+++ oxytcmri/test_oxytcmri.py
@@ -25,9 +25,10 @@ from oxytcmri.models import Subject, Center, MRIExam, MRIVolume, Base, get_cente
 from oxytcmri.utils import get_subject_folder_path, create_tree_structure
 from oxytcmri.usecases.add_clinical_data import (AddClinicalData,
                                                  ClinicalDataRepository,
-                                                 AdditionalClinicalDataRepository, AdditionalClinicalData)
+                                                 AdditionalClinicalDataRepository, AdditionalClinicalData,
+                                                 ClinicalDataDecoder)
 from oxytcmri.infrastructure.clinical_data_repositories import CSVAdditionalClinicalDataRepository, \
-    ExcelClinicalDataRepository, ClinicalDataDecoder
+    ExcelClinicalDataRepository
 
 # The following lines are meant to import the CLI script from the parent directory.
 # See https://www.geeksforgeeks.org/python-import-from-parent-directory/ for more details.
